rename stuff & restructure package to match project mvp
