--- src/pytelliot/tellormon/__init__.py
@@ -1,7 +1,0 @@
-""" tellormon
-
-The tellormon package provides utilities for querying and monitoring the
-TellorX contract, disputing and voting
-
-TBD: Should it maintain a database for easier access to state history
-"""
