+++ src/telliot/cli.py
@@ -1,0 +1,6 @@
+def main() -> None:
+    pass
+
+
+if __name__ == "__main__":
+    main()
