+++ src/telliot/submitter/profitcalc.py
@@ -1,0 +1,4 @@
+""" Profit Calculation
+
+This module provides utilities for determining if transactions will be profitable given a user's configuration settings.
+"""
